requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
captcha==0.3
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
demjson==2.2.4
Django==3.1.2
django-celery-beat==2.0.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-simple-captcha==0.5.13
django-timezone-field==4.0
djangorestframework==3.12.1
docopt==0.6.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
lxml==4.6.1
MarkupSafe==1.1.1
Pillow==7.2.0
prompt-toolkit==3.0.8
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.1
PyMySQL==0.10.1
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
pyvmomi==7.0
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
tyadmin-api-cli==0.2.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
